src/test/java/com/spring/demo/controller/HomeControllerTest.java
package com.spring.demo.controller;

import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.httpBasic;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.context.annotation.Import;
import org.springframework.security.test.context.support.WithMockUser;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;

import com.spring.demo.config.SecurityConfig;
import com.spring.demo.service.TokenService;

@WebMvcTest({HomeController.class, AuthController.class})
@Import({SecurityConfig.class, TokenService.class})
class HomeControllerTest {

    @Autowired
    MockMvc mvc;
    
    @Test
    void secureWhenUnauthenticatedThen401() throws Exception {
        this.mvc.perform(get("/secure"))
        	.andExpect(status().isUnauthorized());
    }

    @Test
    void secureWhenAuthenticatedThenSaysHelloUser() throws Exception {
        MvcResult result = this.mvc.perform(post("/token")
                        .with(httpBasic("amol", "password")))
                .andExpect(status().isOk())
                .andReturn();

        String token = result.getResponse().getContentAsString();

        this.mvc.perform(get("/secure")
                        .header("Authorization", "Bearer " + token))
                .andExpect(content().string("You are visiting a secured endpoint Mr.amol"));
    }

	@Test
	@WithMockUser
    public void rootWithMockUserStatusIsOK() throws Exception {
        this.mvc.perform(get("/")).andExpect(status().isOk());
    }

	@Test
    public void testURLWithMockUserStatusIsOK() throws Exception {
        this.mvc.perform(get("/")).andExpect(content().string("Hello, there!"));
    }
	
}
